move HaveStatusCondition function to fixture
